Extract AgentPrincipal operations into document_config/principal.rs
